Move startup progress out of App into own component and route.
