+++ frontend/tests/fixtures/StartupPage.ts
@@ -6,7 +6,7 @@ export class StartupPage {
   constructor(public readonly mobiFlightPage: MobiFlightPage) {}
 
   async gotoStartupPage() {
-    await this.mobiFlightPage.page.goto("http://localhost:5173/index.html", { waitUntil: "networkidle" });
+    await this.mobiFlightPage.page.goto("http://localhost:5173/start", { waitUntil: "networkidle" });
   }
 
   async setStatusBarUpdate(value: number, text: string) {
